# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Cashflow Minimal API",
    "version": "1.0.0",
    "description": "Read-only Lunch Money ingest + local cashflow controls. Minimal endpoints for GPT Actions."
  },
  "servers": [
    { "url": "https://cashflow.mira.joseai.dev" }
  ],
  "paths": {
...
    "/budgets": {
      "get": {
        "summary": "List budgets for a month",
        "parameters": [
          { "in": "query", "name": "year", "required": true, "schema": { "type": "integer" } },
          { "in": "query", "name": "month", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 12 } }
        ],
        "responses": { "200": { "description": "Array", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BudgetItem" } } } } } }
      }
    },
    "/budgets/{category_id}/{year}-{month}": {
      "put": {
        "summary": "Upsert a monthly budget row",
        "parameters": [
          { "in": "path", "name": "category_id", "required": true, "schema": { "type": "string" } },
          { "in": "path", "name": "year", "required": true, "schema": { "type": "integer" } },
          { "in": "path", "name": "month", "required": true, "schema": { "type": "integer", "minimum": 1, "maximum": 12 } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BudgetUpsert" } } } },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object" } } } } }
      }
    },
...
    "/bills": {
      "get": {
        "summary": "List bills",
        "responses": { "200": { "description": "Array of bills", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Bill" } } } } } }
      },
      "post": {
        "summary": "Create bill",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillCreate" } } } },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "id": { "type": "string" }, "created": { "type": "boolean" } }, "required": ["ok","id","created"] } } } } }
      }
    },
    "/bills/{id}": {
      "patch": {
        "summary": "Patch bill",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillPatch" } } } },
        "responses": { "200": { "description": "OK" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "summary": "Delete bill",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": { "204": { "description": "Deleted" }, "404": { "description": "Not found" } }
      }
    },
    "/bills/{id}/contribute": {
      "post": {
        "summary": "Add contribution to bill ledger",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContribBody" } } } },
        "responses": { "200": { "description": "OK" } }
      }
    },
    "/bills/{id}/mark-paid": {
      "post": {
        "summary": "Mark bill as paid (idempotent)",
        "parameters": [{ "in": "path", "name": "id", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "object", "properties": { "ok": { "type": "boolean" }, "already_paid": { "type": "boolean" }, "period_start": { "type": "string" }, "due": { "type": "string" } }, "required": ["ok","already_paid"] } } } } }
      }
    },
    "/bills/occurrences": {
      "get": {
        "summary": "List bill occurrences with progress",
        "parameters": [
          { "in": "query", "name": "from", "required": true, "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "to", "required": true, "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "due_soon_days", "required": false, "schema": { "type": "integer", "default": 14 } }
        ],
        "responses": { "200": { "description": "Array", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/BillOccurrence" } } } } } }
      }
    },
    "/bills/{id}/ledger": {
      "get": {
        "summary": "Inspect bill ledger in a period",
        "parameters": [
          { "in": "path", "name": "id", "required": true, "schema": { "type": "string" } },
          { "in": "query", "name": "from", "required": true, "schema": { "type": "string", "format": "date" } },
          { "in": "query", "name": "to", "required": true, "schema": { "type": "string", "format": "date" } }
        ],
        "responses": { "200": { "description": "Ledger summary", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BillLedgerResponse" } } } } }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetUpsert": {
        "type": "object",
        "properties": { "amount": { "type": "number" } },
        "required": ["amount"]
      },
      "BudgetItem": {
        "type": "object",
        "properties": {
          "category_id": { "type": "string" },
          "category_name": { "type": "string" },
          "year": { "type": "integer" },
          "month": { "type": "integer" },
          "amount": { "type": "number" }
        }
      },
...
      "Bill": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "amount": { "type": "number" },
          "currency": { "type": "string" },
          "frequency": { "type": "string", "enum": ["weekly","monthly","yearly"] },
          "weekday": { "type": "integer", "nullable": true },
          "day_of_month": { "type": "integer", "nullable": true },
          "start_date": { "type": "string", "format": "date", "nullable": true },
          "end_date": { "type": "string", "format": "date", "nullable": true }
        }
      },
      "BillCreate": { "$ref": "#/components/schemas/Bill" },
      "BillPatch": { "$ref": "#/components/schemas/Bill" },
      "ContribBody": {
        "type": "object",
        "properties": { "amount": { "type": "number" }, "note": { "type": "string" } },
        "required": ["amount"]
      },
      "BillOccurrence": {
        "type": "object",
        "properties": {
          "bill_id": { "type": "string" },
          "name": { "type": "string" },
          "due": { "type": "string", "format": "date" },
          "amount": { "type": "number" },
          "progress_pct": { "type": "number" },
          "paid": { "type": "boolean" },
          "contrib_sum": { "type": "number" },
          "paid_sum": { "type": "number" },
          "days_to_due": { "type": "integer" },
          "due_soon": { "type": "boolean" },
          "priority_score": { "type": "number" }
        }
      },
      "BillLedgerResponse": {
        "type": "object",
        "properties": {
          "bill_id": { "type": "string" },
          "from": { "type": "string" },
          "to": { "type": "string" },
          "contrib_sum": { "type": "number" },
          "paid_sum": { "type": "number" },
          "entries": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "occurred_at": { "type": "string" },
                "amount": { "type": "number" },
                "note": { "type": "string" }
              }
            }
          }
        }
      },
...
    }
  }
}